refactor: snapshot initial zoom
